Refactor schema update into focused modules
